fix: use a forked EntityManager for each transaction (#12)(@corbt)
Currently, a global EntityManager is used for every operation. This causes an error that can be resolved by setting `allowGlobalContext: true`. But that isn't a great fix, because if you ever make an edit to an entity that fails, you'll never be able to make another change because the global context caches the failed change and never releases it.